+++ orgsec-common/src/main/java/com/nomendi6/orgsec/common/service/PrivilegeChecker.java
@@ -62,7 +62,8 @@ public class PrivilegeChecker {
             (operation == PrivilegeOperation.WRITE && resourceAggregatedPrivs.operation == PrivilegeOperation.WRITE) ||
             (operation == PrivilegeOperation.READ &&
                 (resourceAggregatedPrivs.operation == PrivilegeOperation.WRITE ||
-                    resourceAggregatedPrivs.operation == PrivilegeOperation.READ))
+                    resourceAggregatedPrivs.operation == PrivilegeOperation.READ)) ||
+            (operation == PrivilegeOperation.EXECUTE && resourceAggregatedPrivs.operation == PrivilegeOperation.EXECUTE)
         );
     }